Fixed clouds and updated collectables
Clouds: editted the Prefab so it should now work with all clouds we use (specifically Cloud_2)

Gems: now they have a small spin to them. Also there is a HUDCanvas now that keeps track of the gems collected.

HUD: it is a prefab that can be editted. Future use for adding health or something